Move functions out of app_manager/models that are only used externally
